convert string helpers to TS and move to shared
exercises needs part of it
