+++ tutor/src/helpers/string.ts
@@ -1,0 +1,6 @@
+export function assignmentHeaderText(type: string) {
+    if(type === 'external') return 'external assignment';
+    return type;
+}
+
+export * from 'shared/helpers/string'
